Add ApplicationLauncherResourceLoader
