+++ Packages/UGF.Application/Runtime/ApplicationLauncherResourceLoaderBase.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace UGF.Application.Runtime
+{
+    public abstract class ApplicationLauncherResourceLoaderBase : IApplicationLauncherResourceLoader
+    {
+        public abstract Task<IApplicationResources> LoadAsync();
+    }
+}
