+++ Packages/UGF.Application/Runtime/ApplicationLauncherResourcesComponent.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using UnityEngine;
 using Object = UnityEngine.Object;
@@ -13,29 +12,13 @@ namespace UGF.Application.Runtime
 
         public List<Object> Resources { get { return m_resources; } }
 
-        public override async Task<IApplicationResources> LoadAsync()
+        public override Task<IApplicationResources> LoadAsync()
         {
-            var resources = new ApplicationResources();
+            var loader = new ApplicationLauncherResourceLoader();
 
-            for (int i = 0; i < m_resources.Count; i++)
-            {
-                Object resource = m_resources[i];
+            loader.Resources.AddRange(m_resources);
 
-                if (resource == null) throw new ArgumentNullException(nameof(resource), $"Resource not specified at index of the list: '{i}'.");
-
-                if (resource is ApplicationResourceAsset asset)
-                {
-                    object result = await asset.BuildAsync();
-
-                    resources.Add(result);
-                }
-                else
-                {
-                    resources.Add(resource);
-                }
-            }
-
-            return resources;
+            return loader.LoadAsync();
         }
     }
 }
